khepri_fun: Read the "StrT" chunk
This chunk contains a list of concatenated binaries. Instructions can
reference strings inside this "StrT" chunk. Instructions are responsible
for providing the offset in this chunk where to find the string, and the
length of that string.

Here is an example from the `transactions` testsuite. The code calls
`http_util:maybe_add_brackets/2` indirectly. Here is an extract from
this funtion source code at the time of this commit:

```erlang
...
maybe_add_brackets(Addr, true) when is_list(Addr) ->
    case is_ipv6_address(Addr) of
        true ->
            [$[|Addr] ++ "]";
...
```

The code uses the $[ character and the "]" string. They are encoded in
the module's "StrT" chunk as concatenared binaries:

```erlang
<<"[]">> %% <<91, 93>>
```

In its compiled form, the code uses the `bs_create_bin` instruction.
This instruction references the two characters, "[" and "]", in the
"StrT" chunk. For instance, the reference to "]":

```erlang
%% Part of the instruction's arguments:
{atom, string},
_Seg,
_Unit,
_Flags,
{u, 1 = Offset} = _Val,       %% <- The offset into the "StrT" chunk.
{integer, 1 = Length} = _Size %% <- The length of the binary to extract
                              %%    from the "StrT" chunk.
```

It's then easy to call `binary:part/2` to extract the wanted string from
the "StrT" chunk:

```erlang
String = binary:part(StrTChunk, {Offset, Length}).
```

Taking the example from the `bs_create_bin` instruction above:

```erlang
<<"]">> = binary:part(<<"[]">>, {1, 1}).
```

The instruction's arguments above become:

```erlang
%% Part of the instruction's arguments:
{atom, string},
_Seg,
_Unit,
_Flags,
{string, <<"]">>} = _Val,
{integer, 1 = Length} = _Size
```